SOL-133 <컴퓨터 사양에 따라 간헐적으로 발생하는 InputManager 버그 수정>

+++ Assets/Scripts/GameScene/Interact/Generator.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using Photon.Pun;
 using UnityEngine;
 
@@ -19,6 +17,7 @@ public class Generator : MonoBehaviourPun, IInteractableObject
 
     public void Interact(int playerId)
     {
+        if (!photonView.IsMine) return;
         if (installedBatteryCount != 0)
         {
             UninstallBattery(playerId);
@@ -27,6 +26,7 @@ public class Generator : MonoBehaviourPun, IInteractableObject
 
     public void TryInstallBattery(HeldItem heldItem)
     {
+        if (!photonView.IsMine) return;
         if (IsAllBatteryInstalled())
         {
             Debug.Log("Generator : 배터리가 이미 가득 찼습니다.");
@@ -44,30 +44,53 @@ public class Generator : MonoBehaviourPun, IInteractableObject
                                           + transform.up * batteryPositionOffset.y
                                           + transform.forward * batteryPositionOffset.z;
 
-        installedBattery[installedBatteryCount] = heldItem.GetItemObject();
+        GameObject heldItemObject = heldItem.GetItemObject();
+        int viewID = heldItemObject.GetPhotonView().ViewID;
+
+        photonView.RPC("BatteryObjectSync", RpcTarget.All, true, viewID);
         heldItem.ReplaceItem(worldPosition, false);
-        Debug.Log($"Generator : 배터리 장착 성공. 장창된 포지션 : {worldPosition}");
-        IncreaseBatteryCount();
+        photonView.RPC("IncreaseBatteryCount", RpcTarget.All);
         Debug.Log($"Generator : 배터리 장착 성공. 현재 장착된 배터리 갯수 : {installedBatteryCount}");
     }
 
     private void UninstallBattery(int playerId)
     {
-        DecreaseBatteryCount();
+        photonView.RPC("DecreaseBatteryCount", RpcTarget.All);
         GameObject uninstallBatteryObject = installedBattery[installedBatteryCount];
         FarmingObject farmingObject = uninstallBatteryObject.GetComponent<FarmingObject>();
         farmingObject.Interact(playerId);
+        photonView.RPC("BatteryObjectSync", RpcTarget.All, false, 0);
         installedBattery[installedBatteryCount] = null;
         Debug.Log($"Generator : 배터리 회수 성공. 현재 장착된 배터리 갯수 : {installedBatteryCount}");
 
     }
 
+    [PunRPC]
+    private void BatteryObjectSync(bool isInstall, int itemViewID)
+    {
+        if (isInstall)
+        {
+            PhotonView itemPhotonView = PhotonView.Find(itemViewID);
+            if (itemPhotonView == null) return;
+
+            GameObject itemObj = itemPhotonView.gameObject;
+            if (itemObj == null) return;
+            installedBattery[installedBatteryCount] = itemObj;
+        }
+        else
+        {
+            installedBattery[installedBatteryCount] = null;
+        }
+    }
+
+    [PunRPC]
     private void IncreaseBatteryCount()
     {
         installedBatteryCount++;
         batteryPositionOffset += new Vector3(0.4f, 0, 0);
     }
 
+    [PunRPC]
     private void DecreaseBatteryCount()
     {
         installedBatteryCount--;
